requirements.txt
python-docx>=1.1.2
openpyxl>=3.1.2
python-dateutil>=2.9
lxml>=5.2.0
